LayoutTests/fast/viewport/viewport-65.html

<head>
    <title>Check that semi-colon is not a valid separator (no spaces).</title>
    <meta name="viewport" content="width=100;initial-scale=1">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
